+++ src/libs/agent-runtime/bedrock/index.ts
@@ -10,7 +10,7 @@ import { AgentRuntimeErrorType } from '../error';
 import {
   ChatCompetitionOptions,
   ChatStreamPayload,
-  EmbeddingItem,
+  Embeddings,
   EmbeddingsOptions,
   EmbeddingsPayload,
   ModelProvider,
@@ -56,11 +56,15 @@ export class LobeBedrockAI implements LobeRuntimeAI {
 
     return this.invokeClaudeModel(payload, options);
   }
-
-  async embeddings(
-    payload: EmbeddingsPayload,
-    options?: EmbeddingsOptions,
-  ): Promise<EmbeddingItem[]> {
+  /**
+   * Supports the Amazon Titan Text models series.
+   * Cohere Embed models are not supported
+   * because the current text size per request
+   * exceeds the maximum 2048 characters limit
+   * for a single request for this series of models.
+   * [bedrock embed guide] https://docs.aws.amazon.com/bedrock/latest/userguide/model-parameters-embed.html
+   */
+  async embeddings(payload: EmbeddingsPayload, options?: EmbeddingsOptions): Promise<Embeddings[]> {
     const input = Array.isArray(payload.input) ? payload.input : [payload.input];
     const promises = input.map((inputText: string, index: number) =>
       this.invokeEmbeddingModel(
@@ -79,7 +83,7 @@ export class LobeBedrockAI implements LobeRuntimeAI {
   private invokeEmbeddingModel = async (
     payload: EmbeddingsPayload,
     options?: EmbeddingsOptions,
-  ): Promise<EmbeddingItem> => {
+  ): Promise<Embeddings> => {
     const command = new InvokeModelCommand({
       accept: 'application/json',
       body: JSON.stringify({
@@ -93,11 +97,7 @@ export class LobeBedrockAI implements LobeRuntimeAI {
     try {
       const res = await this.client.send(command, { abortSignal: options?.signal });
       const responseBody = JSON.parse(new TextDecoder().decode(res.body));
-      return {
-        embedding: responseBody.embedding,
-        index: payload.index as number,
-        object: 'embedding',
-      };
+      return responseBody.embedding;
     } catch (e) {
       const err = e as Error & { $metadata: any };
       throw AgentRuntimeError.chat({